API/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
pymysql==1.1.0
python-dotenv==1.0.1
pydantic==2.6.4
psycopg2